refactor(cell): split base_messages.rs (1797 lines) into directory module
Extract the monolithic base_messages handler into submodules along
variant-family seams:

- mod.rs: dispatch function with small handlers inline
- player_init.rs: InitPlayerState handler (mission/ability/bandolier restore)
- bandolier.rs: UpdateBandolierItem + SyncBandolierItems handlers
- tests/: test suite split by handler family (general, bandolier_update,
  bandolier_sync)

All 290 original comment lines preserved (308 total with new module docs).
