+++ crates/services/src/cell/service/base_messages/tests/mod.rs
@@ -1,0 +1,8 @@
+use super::*;
+use crate::cell::space_manager::SpaceManager;
+use cimmeria_entity::cell_entity::BandolierItem;
+
+
+mod bandolier_sync;
+mod bandolier_update;
+mod general;
